+++ data/types/upload/upload_test.go
@@ -1,11 +1,11 @@
 package upload_test
 
 import (
+	"sort"
+
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/ginkgo/extensions/table"
 	. "github.com/onsi/gomega"
-
-	"sort"
 
 	dataNormalizer "github.com/tidepool-org/platform/data/normalizer"
 	dataTypes "github.com/tidepool-org/platform/data/types"