+++ dexcom/time_test.go
@@ -1,10 +1,10 @@
 package dexcom_test
 
 import (
+	"fmt"
+
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-
-	"fmt"
 
 	"github.com/tidepool-org/platform/dexcom"
 	dexcomTest "github.com/tidepool-org/platform/dexcom/test"