+++ user/store/mongo/config_test.go
@@ -1,10 +1,10 @@
 package mongo_test
 
 import (
+	"time"
+
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-
-	"time"
 
 	"github.com/tidepool-org/platform/config/test"
 	"github.com/tidepool-org/platform/pointer"