+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## HEAD
 
+* Fix unused tests
+* Fix import ordering due to Golang tool changes
+* Update dependencies
+* Update to Go 1.11.1
+* Add user id to ArchiveDeviceDataUsingHashesFromDataSet selector to properly use indexes
+* Fix data set id validation to allow Jellyfish id format
+
 ## v1.30.0
 
 * Fix Dexcom API v2 edge cases